[feature] Created a validator that has a promise wrapper and is extremely easy and clear to use.
So you just import the mail_validator.js file, which returns an object with two things, a validator function and a schema table.
It is extremely easy to use and structured in a way you dont have to think about the implementation to use it.
The validator is promised based, so if you call it, it resolves a `value` or rejects an `error`.
slackName: Joe.Attah
Team Fierce.